Extracted raster params frame
