+++ gui/acquisition/__init__.py
@@ -1,3 +1,4 @@
 from .vector_params import VectorParamsFrame
 from .multicol_params import MultiColParamsFrame
 from .characterize_params import CharacterizeParamsFrame
+from .raster_params import RasterParamsFrame
