+++ gui/control_main.py
@@ -1,4 +1,3 @@
-import _thread
 import functools
 import logging
 import math
@@ -52,6 +51,7 @@ from gui.acquisition import (
     VectorParamsFrame,
     MultiColParamsFrame,
     CharacterizeParamsFrame,
+    RasterParamsFrame,
 )
 from gui.raster import RasterCell, RasterGroup
 from QPeriodicTable import QPeriodicTable
@@ -200,7 +200,6 @@ class ControlMain(QtWidgets.QMainWindow):
 
         self.beamSize_pv = PV(daq_utils.beamlineComm + "size_mode")
         self.energy_pv = PV(daq_utils.motor_dict["energy"] + ".RBV")
-        self.rasterStepDefs = {"Coarse": 20.0, "Fine": 10.0, "VFine": 5.0}
 
         # Timer that waits for a second before calling raddose 3d
         # This is to prevent multiple calls when transmission textbox is changing
@@ -574,22 +573,6 @@ class ControlMain(QtWidgets.QMainWindow):
         hBoxColParams22.insertSpacing(5, 100)
         hBoxColParams22.addWidget(beamsizeLabel)
         hBoxColParams22.addWidget(self.beamsizeComboBox)
-        hBoxColParams4 = QtWidgets.QHBoxLayout()
-        colBeamWLabel = QtWidgets.QLabel("Beam Width:")
-        colBeamWLabel.setFixedWidth(140)
-        colBeamWLabel.setAlignment(QtCore.Qt.AlignCenter)
-        self.beamWidth_ledit = QtWidgets.QLineEdit()
-        self.beamWidth_ledit.setFixedWidth(60)
-        colBeamHLabel = QtWidgets.QLabel("Beam Height:")
-        colBeamHLabel.setFixedWidth(140)
-        colBeamHLabel.setAlignment(QtCore.Qt.AlignCenter)
-        self.beamHeight_ledit = QtWidgets.QLineEdit()
-        self.beamHeight_ledit.setFixedWidth(60)
-        hBoxColParams4.addWidget(colBeamWLabel)
-        hBoxColParams4.addWidget(self.beamWidth_ledit)
-        hBoxColParams4.addWidget(colBeamHLabel)
-        hBoxColParams4.addWidget(self.beamHeight_ledit)
-        hBoxColParams5 = QtWidgets.QHBoxLayout()
         colResoLabel = QtWidgets.QLabel("Edge Resolution:")
         colResoLabel.setAlignment(QtCore.Qt.AlignCenter)
         self.resolution_ledit = QtWidgets.QLineEdit()
@@ -712,61 +695,7 @@ class ControlMain(QtWidgets.QMainWindow):
         self.hBoxProcessingLayout1.addWidget(self.fastEPCheckBox)
         self.hBoxProcessingLayout1.addWidget(self.dimpleCheckBox)
         self.processingOptionsFrame.setLayout(self.hBoxProcessingLayout1)
-        self.rasterParamsFrame = QFrame()
-        self.vBoxRasterParams = QtWidgets.QVBoxLayout()
-        self.hBoxRasterLayout1 = QtWidgets.QHBoxLayout()
-        self.hBoxRasterLayout1.setAlignment(QtCore.Qt.AlignLeft)
-        self.hBoxRasterLayout2 = QtWidgets.QHBoxLayout()
-        self.hBoxRasterLayout2.setAlignment(QtCore.Qt.AlignLeft)
-        rasterStepLabel = QtWidgets.QLabel("Raster Step")
-        rasterStepLabel.setFixedWidth(110)
-        self.rasterStepEdit = QtWidgets.QLineEdit(str(self.rasterStepDefs["Coarse"]))
-        self.rasterStepEdit.textChanged[str].connect(self.rasterStepChanged)
-        self.rasterStepEdit.setFixedWidth(60)
-        self.rasterGrainRadioGroup = QtWidgets.QButtonGroup()
-        self.rasterGrainCoarseRadio = QtWidgets.QRadioButton("Coarse")
-        self.rasterGrainCoarseRadio.setChecked(False)
-        self.rasterGrainCoarseRadio.toggled.connect(
-            functools.partial(self.rasterGrainToggledCB, "Coarse")
-        )
-        self.rasterGrainRadioGroup.addButton(self.rasterGrainCoarseRadio)
-        self.rasterGrainFineRadio = QtWidgets.QRadioButton("Fine")
-        self.rasterGrainFineRadio.setChecked(False)
-        self.rasterGrainFineRadio.toggled.connect(
-            functools.partial(self.rasterGrainToggledCB, "Fine")
-        )
-        self.rasterGrainRadioGroup.addButton(self.rasterGrainFineRadio)
-        self.rasterGrainVFineRadio = QtWidgets.QRadioButton("VFine")
-        self.rasterGrainVFineRadio.setChecked(False)
-        self.rasterGrainVFineRadio.toggled.connect(
-            functools.partial(self.rasterGrainToggledCB, "VFine")
-        )
-        self.rasterGrainRadioGroup.addButton(self.rasterGrainVFineRadio)
-        self.rasterGrainCustomRadio = QtWidgets.QRadioButton("Custom")
-        self.rasterGrainCustomRadio.setChecked(True)
-        self.rasterGrainCustomRadio.toggled.connect(
-            functools.partial(self.rasterGrainToggledCB, "Custom")
-        )
-        self.rasterGrainRadioGroup.addButton(self.rasterGrainCustomRadio)
-        rasterEvalLabel = QtWidgets.QLabel("Raster\nEvaluate By:")
-        rasterEvalOptionList = ["Spot Count", "Resolution", "Intensity"]
-        self.rasterEvalComboBox = QtWidgets.QComboBox(self)
-        self.rasterEvalComboBox.addItems(rasterEvalOptionList)
-        self.rasterEvalComboBox.setCurrentIndex(
-            db_lib.beamlineInfo(daq_utils.beamline, "rasterScoreFlag")["index"]
-        )
-        self.rasterEvalComboBox.activated[str].connect(self.rasterEvalComboActivatedCB)
-        self.hBoxRasterLayout1.addWidget(rasterStepLabel)
-        self.hBoxRasterLayout1.addWidget(self.rasterStepEdit)
-        self.hBoxRasterLayout1.addWidget(self.rasterGrainCoarseRadio)
-        self.hBoxRasterLayout1.addWidget(self.rasterGrainFineRadio)
-        self.hBoxRasterLayout1.addWidget(self.rasterGrainVFineRadio)
-        self.hBoxRasterLayout1.addWidget(self.rasterGrainCustomRadio)
-        self.hBoxRasterLayout1.addWidget(rasterEvalLabel)
-        self.hBoxRasterLayout1.addWidget(self.rasterEvalComboBox)
-        self.vBoxRasterParams.addLayout(self.hBoxRasterLayout1)
-        self.vBoxRasterParams.addLayout(self.hBoxRasterLayout2)
-        self.rasterParamsFrame.setLayout(self.vBoxRasterParams)
+        self.rasterParamsFrame = RasterParamsFrame(self)
         self.multiColParamsFrame = MultiColParamsFrame(self)
         self.characterizeParamsFrame = CharacterizeParamsFrame(self)
         self.vectorParamsFrame = VectorParamsFrame(self)
@@ -1389,13 +1318,6 @@ class ControlMain(QtWidgets.QMainWindow):
             self.fastEPCheckBox.setEnabled(False)
             self.dimpleCheckBox.setEnabled(False)
             self.xia2CheckBox.setEnabled(False)
-
-    def rasterGrainToggledCB(self, identifier):
-        if identifier == "Coarse" or identifier == "Fine" or identifier == "VFine":
-            cellSize = self.rasterStepDefs[identifier]
-            self.rasterStepEdit.setText(str(cellSize))
-            self.beamWidth_ledit.setText(str(cellSize))
-            self.beamHeight_ledit.setText(str(cellSize))
 
     def vidActionToggledCB(self):
         if len(self.rasterList) > 0:
@@ -2223,10 +2145,6 @@ class ControlMain(QtWidgets.QMainWindow):
         else:
             pass
 
-    def rasterStepChanged(self, text):
-        self.beamWidth_ledit.setText(text)
-        self.beamHeight_ledit.setText(text)
-
     def updateVectorLengthAndSpeed(self):
         x_vec_end = self.vectorEnd["coords"]["x"]
         y_vec_end = self.vectorEnd["coords"]["y"]
@@ -2511,15 +2429,6 @@ class ControlMain(QtWidgets.QMainWindow):
         else:
             self.protoOtherRadio.setChecked(True)
         self.totalExpChanged("")
-
-    def rasterEvalComboActivatedCB(self, text):
-        db_lib.beamlineInfo(
-            daq_utils.beamline,
-            "rasterScoreFlag",
-            info_dict={"index": self.rasterEvalComboBox.findText(str(text))},
-        )
-        if self.currentRasterCellList != []:
-            self.reFillPolyRaster()
 
     def popBaseDirectoryDialogCB(self):
         fname = QtWidgets.QFileDialog.getExistingDirectory(
@@ -2829,8 +2738,8 @@ class ControlMain(QtWidgets.QMainWindow):
         raster_h = int(self.polyBoundingRect.height())
         center_x = int(self.polyBoundingRect.center().x())
         center_y = int(self.polyBoundingRect.center().y())
-        stepsizeXPix = self.screenXmicrons2pixels(float(self.rasterStepEdit.text()))
-        stepsizeYPix = self.screenYmicrons2pixels(float(self.rasterStepEdit.text()))
+        stepsizeXPix = self.screenXmicrons2pixels(self.rasterParamsFrame.rasterStep)
+        stepsizeYPix = self.screenYmicrons2pixels(self.rasterParamsFrame.rasterStep)
         self.click_positions = []
         self.definePolyRaster(
             raster_w, raster_h, stepsizeXPix, stepsizeYPix, center_x, center_y
@@ -2905,7 +2814,7 @@ class ControlMain(QtWidgets.QMainWindow):
         spotLineCounter = 0
         cellIndex = 0
         rowStartIndex = 0
-        rasterEvalOption = str(self.rasterEvalComboBox.currentText())
+        rasterEvalOption = str(self.rasterParamsFrame.rasterEvalComboBox.currentText())
         lenX = abs(
             rasterDef["rowDefs"][0]["end"]["x"] - rasterDef["rowDefs"][0]["start"]["x"]
         )  # ugly for tile flip/noflip
@@ -3048,7 +2957,7 @@ class ControlMain(QtWidgets.QMainWindow):
             logger.error(f"Exception while writing raster result: {e}")
 
     def reFillPolyRaster(self):
-        rasterEvalOption = str(self.rasterEvalComboBox.currentText())
+        rasterEvalOption = str(self.rasterParamsFrame.rasterEvalComboBox.currentText())
         for i in range(len(self.rasterList)):
             if self.rasterList[i] != None:
                 currentRasterGroup = self.rasterList[i]["graphicsItem"]
@@ -3239,7 +3148,7 @@ class ControlMain(QtWidgets.QMainWindow):
     ):  # all come in as pixels, raster_w and raster_h are bounding box of drawn graphic
         # raster status - 0=nothing done, 1=run, 2=displayed
         stepTime = float(self.exp_time_ledit.text())
-        stepsize = float(self.rasterStepEdit.text())
+        stepsize = self.rasterParamsFrame.rasterStep
         if (stepsize / 1000.0) / stepTime > 2.0:
             self.popupServerMessage(
                 "Stage speed exceeded. Increase exposure time, or decrease step size. Limit is 2mm/s."
@@ -3248,8 +3157,7 @@ class ControlMain(QtWidgets.QMainWindow):
             return
 
         try:
-            beamWidth = float(self.beamWidth_ledit.text())
-            beamHeight = float(self.beamHeight_ledit.text())
+            beamWidth = beamHeight = self.rasterParamsFrame.rasterStep
         except ValueError:
             logger.error("bad value for beam width or beam height")
             self.popupServerMessage("bad value for beam width or beam height")
@@ -3615,8 +3523,8 @@ class ControlMain(QtWidgets.QMainWindow):
                     self.dataPathGB.file_numstart_ledit.text()
                 )
             reqObj["attenuation"] = float(self.transmission_ledit.text())
-            reqObj["slit_width"] = float(self.beamWidth_ledit.text())
-            reqObj["slit_height"] = float(self.beamHeight_ledit.text())
+            reqObj["slit_width"] = self.rasterParamsFrame.rasterStep
+            reqObj["slit_height"] = self.rasterParamsFrame.rasterStep
             reqObj["energy"] = float(self.energy_ledit.text())
             wave = daq_utils.energy2wave(float(self.energy_ledit.text()), digits=6)
             reqObj["wavelength"] = wave
@@ -3888,8 +3796,8 @@ class ControlMain(QtWidgets.QMainWindow):
                             self.dataPathGB.file_numstart_ledit.text()
                         )
                         reqObj["attenuation"] = float(self.transmission_ledit.text())
-                        reqObj["slit_width"] = float(self.beamWidth_ledit.text())
-                        reqObj["slit_height"] = float(self.beamHeight_ledit.text())
+                        reqObj["slit_width"] = self.rasterParamsFrame.rasterStep
+                        reqObj["slit_height"] = self.rasterParamsFrame.rasterStep
                         reqObj["energy"] = float(self.energy_ledit.text())
                         wave = daq_utils.energy2wave(
                             float(self.energy_ledit.text()), digits=6
@@ -4019,8 +3927,8 @@ class ControlMain(QtWidgets.QMainWindow):
                     reqObj["dimple"] = self.dimpleCheckBox.isChecked()
                 reqObj["xia2"] = self.xia2CheckBox.isChecked()
                 reqObj["attenuation"] = float(self.transmission_ledit.text())
-                reqObj["slit_width"] = float(self.beamWidth_ledit.text())
-                reqObj["slit_height"] = float(self.beamHeight_ledit.text())
+                reqObj["slit_width"] = self.rasterParamsFrame.rasterStep
+                reqObj["slit_height"] = self.rasterParamsFrame.rasterStep
                 reqObj["energy"] = float(self.energy_ledit.text())
             except ValueError:
                 message = "Please ensure that all boxes that expect numerical values have numbers in them"
@@ -4041,15 +3949,15 @@ class ControlMain(QtWidgets.QMainWindow):
                 logger.error("set dist to %s in exception handler 1" % new_distance)
                 reqObj["detDist"] = new_distance
             if reqObj["protocol"] == "multiCol" or reqObj["protocol"] == "multiColQ":
-                reqObj["gridStep"] = float(self.rasterStepEdit.text())
+                reqObj["gridStep"] = self.rasterParamsFrame.rasterStep
                 reqObj["diffCutoff"] = float(
                     self.multiColParamsFrame.multiColCutoffEdit.text()
                 )
             if reqObj["protocol"] == "rasterScreen":
-                reqObj["gridStep"] = float(self.rasterStepEdit.text())
+                reqObj["gridStep"] = self.rasterParamsFrame.rasterStep
             if rasterDef != None:
                 reqObj["rasterDef"] = rasterDef
-                reqObj["gridStep"] = float(self.rasterStepEdit.text())
+                reqObj["gridStep"] = self.rasterParamsFrame.rasterStep
             if reqObj["protocol"] == "characterize" or reqObj["protocol"] == "ednaCol":
                 reqObj[
                     "characterizationParams"
@@ -4417,8 +4325,8 @@ class ControlMain(QtWidgets.QMainWindow):
             }
         )
         self.dataPathGB.setFileNumstart_ledit(str(reqObj["file_number_start"]))
-        self.beamWidth_ledit.setText(str(reqObj["slit_width"]))
-        self.beamHeight_ledit.setText(str(reqObj["slit_height"]))
+        self.rasterParamsFrame.rasterStep = reqObj["slit_width"]
+        # self.rasterParamsFrame.beamHeight = reqObj["slit_height"]
         if "fastDP" in reqObj:
             self.staffScreenDialog.fastDPCheckBox.setChecked(
                 (reqObj["fastDP"] or reqObj["fastEP"] or reqObj["dimple"])
@@ -4470,15 +4378,8 @@ class ControlMain(QtWidgets.QMainWindow):
                                 f"Master HDF5 file {firstFilename} could not be validated",
                                 QtWidgets.QMessageBox.Ok,
                             )
-        self.rasterStepEdit.setText(str(reqObj["gridStep"]))
-        if reqObj["gridStep"] == self.rasterStepDefs["Coarse"]:
-            self.rasterGrainCoarseRadio.setChecked(True)
-        elif reqObj["gridStep"] == self.rasterStepDefs["Fine"]:
-            self.rasterGrainFineRadio.setChecked(True)
-        elif reqObj["gridStep"] == self.rasterStepDefs["VFine"]:
-            self.rasterGrainVFineRadio.setChecked(True)
-        else:
-            self.rasterGrainCustomRadio.setChecked(True)
+        self.rasterParamsFrame.setGridStep(reqObj["gridStep"])
+
         rasterStep = int(reqObj["gridStep"])
         if not self.hideRastersCheckBox.isChecked() and (
             reqObj["protocol"] in ("raster", "stepRaster", "multiCol")
